+++ vector/src/main/java/im/vector/riotx/core/utils/ExternalApplicationsUtil.kt
@@ -26,6 +26,7 @@ import android.net.Uri
 import android.os.Build
 import android.provider.Browser
 import android.provider.MediaStore
+import android.widget.Toast
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.browser.customtabs.CustomTabsSession
 import androidx.core.content.ContextCompat
@@ -33,8 +34,10 @@ import androidx.core.content.FileProvider
 import androidx.fragment.app.Fragment
 import im.vector.riotx.BuildConfig
 import im.vector.riotx.R
+import im.vector.riotx.features.notifications.NotificationUtils
 import okio.buffer
 import okio.sink
+import okio.source
 import timber.log.Timber
 import java.io.File
 import java.text.SimpleDateFormat
@@ -296,7 +299,7 @@ fun shareMedia(context: Context, file: File, mediaMimeType: String?) {
     }
 }
 
-fun saveMedia(context: Context, file: File, title: String, mediaMimeType: String?): Boolean {
+fun saveMedia(context: Context, file: File, title: String, mediaMimeType: String?, notificationUtils: NotificationUtils) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
         val externalContentUri: Uri
         val values = ContentValues()
@@ -334,21 +337,34 @@ fun saveMedia(context: Context, file: File, title: String, mediaMimeType: String
                 values.put(MediaStore.Downloads.DATE_TAKEN, System.currentTimeMillis())
             }
         }
-        context.contentResolver.insert(externalContentUri, values)?.let { uri ->
-            context.contentResolver.openOutputStream(uri)?.use { outputStream ->
-                outputStream.sink().buffer().write(file.inputStream().use { it.readBytes() })
-                return true
+        val uri = context.contentResolver.insert(externalContentUri, values)
+        if (uri == null) {
+            Toast.makeText(context, R.string.error_saving_media_file, Toast.LENGTH_LONG).show()
+        } else {
+            val source = file.inputStream().source().buffer()
+            context.contentResolver.openOutputStream(uri)?.sink()?.buffer()?.let { sink ->
+                source.use { input ->
+                    sink.use { output ->
+                        output.writeAll(input)
+                    }
+                }
+            }
+            notificationUtils.buildDownloadFileNotification(
+                    uri,
+                    title,
+                    mediaMimeType ?: "application/octet-stream"
+            ).let { notification ->
+                notificationUtils.showNotificationMessage("DL", uri.hashCode(), notification)
             }
         }
+        // TODO add notification?
     } else {
         @Suppress("DEPRECATION")
         Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE).also { mediaScanIntent ->
             mediaScanIntent.data = Uri.fromFile(file)
             context.sendBroadcast(mediaScanIntent)
         }
-        return true
     }
-    return false
 }
 
 /**